+++ client/src/components/Copyright/index.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import Link from '@material-ui/core/Link';
+import Typography from '@material-ui/core/Typography';
+
+function Copyright() {
+    return (
+        <Typography variant="body2" color="textSecondary" align="center">
+            {'Copyright © '}
+            <Link color="inherit" href="https://material-ui.com/">
+                Camp Quest Kansas City
+      </Link>{' '}
+            {new Date().getFullYear()}
+            {'.'}
+        </Typography>
+    );
+}
+
+export default Copyright;
