+++ client/src/pages/Login.js
@@ -11,8 +11,7 @@ import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
 import Container from '@material-ui/core/Container';
 import CQKC_Logo_wBorder from "../images/CQKC_Logo_wBorder.png"
-import "../pages/style.css"
-
+import Copyright from "../components/Copyright"
 
 const useStyles = makeStyles((theme) => ({
     paper: {
@@ -34,6 +33,8 @@ const useStyles = makeStyles((theme) => ({
     },
 }));
 
+
+
 export default function SignIn() {
     const classes = useStyles();
 
@@ -44,51 +45,44 @@ export default function SignIn() {
                 <div>
                     <img className="logo_wBorder" src={CQKC_Logo_wBorder} alt="CQKC Logo" />
                 </div>
-                <Typography component="h1" variant="h5">
 
+                <Typography component="h1" variant="h5">
                     Sign in
-        </Typography>
+                 </Typography>
+
                 <form className={classes.form} noValidate>
                     <TextField
                         variant="outlined"
                         margin="normal"
-                        required
-                        fullWidth
-                        id="email"
+                        required fullWidth id="email"
                         label="Email Address"
                         name="email"
                         autoComplete="email"
                         autoFocus
                     />
+
                     <TextField
                         variant="outlined"
                         margin="normal"
-                        required
-                        fullWidth
-                        name="password"
+                        required fullWidth name="password"
                         label="Password"
                         type="password"
                         id="password"
                         autoComplete="current-password"
                     />
+
                     <FormControlLabel
                         control={<Checkbox value="remember" color="primary" />}
                         label="Remember me"
                     />
-                    <Button
-                        type="submit"
-                        fullWidth
-                        variant="contained"
-                        color="primary"
-                        className={classes.submit}
-                    >
+
+                    <Button type="submit" fullWidth variant="contained" color="primary" className={classes.submit}>
                         Sign In
-          </Button>
+                    </Button>
+
                     <Grid container>
                         <Grid item xs>
-                            <Link href="#" variant="body2">
-                                Forgot password?
-              </Link>
+                            <Link href="#" variant="body2">Forgot password?</Link>
                         </Grid>
                         <Grid item>
                             <Link href="#" variant="body2">
@@ -98,22 +92,15 @@ export default function SignIn() {
                     </Grid>
                 </form>
             </div>
+
             <Box mt={8}>
                 <Copyright />
             </Box>
+
         </Container>
+
+
     );
 }
 
-function Copyright() {
-    return (
-        <Typography variant="body2" color="textSecondary" align="center">
-            {'Copyright © '}
-            <Link color="inherit" href="https://material-ui.com/">
-                Camp Quest Kansas City
-      </Link>{' '}
-            {new Date().getFullYear()}
-            {'.'}
-        </Typography>
-    );
-}
+
